Split parser classes into .class.php + trait per naming convention
